Time: 0 ms (100%), Space: 8.3 MB (46.07%) - LeetHub

+++ 0190-reverse-bits/0190-reverse-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    uint32_t reverseBits(uint32_t n) {
-        uint32_t result = 0;
+    long long int reverseBits(uint32_t n) {
+        long long int result = 0;
         
         for(int i = 0; i < 32; i++) {
             result <<= 1;
